[#23] 메인 페이지 CSS 수정, 컴포넌트 분리
