+++ src/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import Footer from './page/common/Footer';
 import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
