--- src/page/common/FooterView.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const FooterDiv = styled.div`
-  height: 154px;
-  width: 1440px;
-  border-top: 1px solid;
-  border-top-color: #868e96;
-`;
